refactor: 유효성 검사 함수 한 곳에 모아둔 곳으로부터 불러오게 해라
